refactor: split assignment matrix modules
