+++ docs/plans/codebase-maintenance-roadmap.md
@@ -41,6 +41,7 @@ This roadmap turns the current maintenance audit into concrete, sequenced backlo
 - 2026-05-16: P3-03 completed by splitting `src/utils/flex-folders/folders.ts` from `1631` lines into a `1` line compatibility entrypoint plus `src/utils/flex-folders/folder-creation/` modules for orchestration, dryhire creation and cleanup, tourdate creation, standard job creation, commercial extras, shared helpers, and shared creation types. The new creation boundary also uses a Flex-folder-specific job type instead of preserving the broad legacy `any`. Local validation passed with `npm run typecheck`, targeted Flex folder Vitest coverage, `npm run lint`, `npm run test:critical`, `npm run test:run`, `npm run build`, `npm run cap:sync`, `npm run test:e2e`, and `git diff --check`.
 - 2026-05-16: P3-04 completed by adding `src/features/technical-tools/` modules for shared Consumos power calculations, PDU recommendations, power/default/override payload builders, report upload/task-completion wrappers, Pesos weight calculations, and reusable power table controls. Sound/lights/video Consumos pages and sound/video Pesos pages now share the core calculation and persistence paths while preserving department-specific labels, component catalogs, PDF titles, and upload categories. Focused coverage includes power math, mixed-light apparent power, payload shape, upload failure versus non-fatal task-completion failure, weight totals, rigging-point formatting, and shared control labels. Local validation passed with `npm run typecheck`, targeted technical-tool and power-summary Vitest coverage, `npm run lint`, `npm run test:critical`, `npm run test:run`, `npm run build`, `npm run cap:sync`, `npm run test:e2e`, and `git diff --check`.
 - 2026-05-17: P3-05 completed by splitting `src/features/festival-management/useFestivalManagementVm.ts` from `1354` lines into a `222` line composition hook plus feature-local query services, command wrappers, derived selectors, and focused hooks for job data, documents, Flex controls, map previews, print actions, administrative commands, and WhatsApp actions. Review follow-up moved job/document state onto TanStack Query, made festival date derivation and Flex date payloads `Europe/Madrid` aware, and added upload cleanup when job document metadata persistence fails. Focused coverage now includes selector behavior for stage/date/rider/Flex/WhatsApp derivations and Playwright smoke coverage for the festival management overview plus the public blank artist requirements form. Local validation passed with `npm run typecheck`, focused Vitest coverage, focused Playwright coverage, `npm run lint`, `npm run test:critical`, `npm run test:run`, `npm run build`, `npm run cap:sync`, `npm run test:e2e`, and `git diff --check`.
+- 2026-05-17: P3-06 completed by splitting the assignment matrix route and render path into focused page controls/reminder helpers, scroll virtualization state, technician ordering/ranking state, matrix dialogs, and cell-level removal/dialog/tooltip modules. The largest matrix files are now under the roadmap limit: `src/pages/JobAssignmentMatrix.tsx` is `697` lines, `src/components/matrix/OptimizedAssignmentMatrix.tsx` is `595` lines, `src/components/matrix/optimized-assignment-matrix/OptimizedAssignmentMatrixView.tsx` is `434` lines, and `src/components/matrix/OptimizedMatrixCell.tsx` is `578` lines. Focused coverage includes the matrix page/unit suites, optimized matrix view/cell suites, optimized matrix data hook coverage, and Playwright smoke coverage for assignment matrix plus staffing recommendations. Local validation passed with `npm run typecheck`, focused Vitest coverage, focused Playwright coverage, `npm run lint`, `npm run test:critical`, `npm run test:run`, `npm run build`, `npm run cap:sync`, `npm run test:e2e`, and `git diff --check`.
 
 ## Phase 0: Safety and Type Baseline
